requirements.txt
asgiref==3.2.10
Django==3.1.2
django-allauth==0.42.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-multiselectfield==0.1.12
oauthlib==3.1.0
Pillow==7.2.0
python3-openid==3.2.0
pytz==2020.1
requests-oauthlib==1.3.0
sqlparse==0.3.1
stripe==2.55.0
